requirements.txt
spacy==3.1.2
ibm_cloud_sdk_core==3.11.3
numpy==1.20.3
Flask==2.0.1
textdescriptives==1.0.3
ibm_watson==5.2.3
syllables==1.0.3
Flask_Cors==3.0.10
regex==2021.8.28
beautifulsoup4==4.10.0
python-dotenv==0.19.0
gunicorn~=20.1.0
click~=7.1.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl

